Remove examples for draft.js 0.9.0 (#2358)
Summary:
The API for draft 9 and in general that version is no longer supported. We don't need these examples. Moreover, I can see them potentially confusing newcomers. Next up, we should look at the examples for v0.10 and decide if they need to be updated for v0.11.
Pull Request resolved: https://github.com/facebook/draft-js/pull/2358

Reviewed By: claudiopro

Differential Revision: D20542820

Pulled By: mrkev

fbshipit-source-id: 6b538dd3c2a2fe3f76532f16e1de066d68547d81